Separate SLT Summary and Environment from bcase class implementation
